Inline utility methods
