+++ instrumentation/apache-httpclient/apache-httpclient-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v5_0/ApacheHttpClientHttpAttributesGetter.java
@@ -6,27 +6,66 @@
 package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v5_0;
 
 import io.opentelemetry.instrumentation.api.instrumenter.http.HttpClientAttributesGetter;
+import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.annotation.Nullable;
+import org.apache.hc.core5.http.Header;
 import org.apache.hc.core5.http.HttpRequest;
 import org.apache.hc.core5.http.HttpResponse;
+import org.apache.hc.core5.http.MessageHeaders;
+import org.apache.hc.core5.http.ProtocolVersion;
+import org.apache.hc.core5.net.URIAuthority;
 
 final class ApacheHttpClientHttpAttributesGetter
     implements HttpClientAttributesGetter<HttpRequest, HttpResponse> {
+  private static final Logger logger =
+      Logger.getLogger(ApacheHttpClientHttpAttributesGetter.class.getName());
 
   @Override
   public String method(HttpRequest request) {
-    return ApacheHttpClientUtils.getMethod(request);
+    return request.getMethod();
   }
 
   @Override
   public String url(HttpRequest request) {
-    return ApacheHttpClientUtils.getUrl(request);
+    // similar to org.apache.hc.core5.http.message.BasicHttpRequest.getUri()
+    // not calling getUri() to avoid unnecessary conversion
+    StringBuilder url = new StringBuilder();
+    URIAuthority authority = request.getAuthority();
+    if (authority != null) {
+      String scheme = request.getScheme();
+      if (scheme != null) {
+        url.append(scheme);
+        url.append("://");
+      } else {
+        url.append("http://");
+      }
+      url.append(authority.getHostName());
+      int port = authority.getPort();
+      if (port >= 0) {
+        url.append(":");
+        url.append(port);
+      }
+    }
+    String path = request.getPath();
+    if (path != null) {
+      if (url.length() > 0 && !path.startsWith("/")) {
+        url.append("/");
+      }
+      url.append(path);
+    } else {
+      url.append("/");
+    }
+    return url.toString();
   }
 
   @Override
   public List<String> requestHeader(HttpRequest request, String name) {
-    return ApacheHttpClientUtils.getHeader(request, name);
+    return getHeader(request, name);
   }
 
   @Override
@@ -44,13 +83,33 @@ final class ApacheHttpClientHttpAttributesGetter
 
   @Override
   public Integer statusCode(HttpRequest request, HttpResponse response) {
-    return ApacheHttpClientUtils.getStatusCode(response);
+    return response.getCode();
   }
 
   @Override
   @Nullable
   public String flavor(HttpRequest request, @Nullable HttpResponse response) {
-    return ApacheHttpClientUtils.getFlavor(request, response);
+    ProtocolVersion protocolVersion = getVersion(request, response);
+    if (protocolVersion == null) {
+      return null;
+    }
+    String protocol = protocolVersion.getProtocol();
+    if (!protocol.equals("HTTP")) {
+      return null;
+    }
+    int major = protocolVersion.getMajor();
+    int minor = protocolVersion.getMinor();
+    if (major == 1 && minor == 0) {
+      return SemanticAttributes.HttpFlavorValues.HTTP_1_0;
+    }
+    if (major == 1 && minor == 1) {
+      return SemanticAttributes.HttpFlavorValues.HTTP_1_1;
+    }
+    if (major == 2 && minor == 0) {
+      return SemanticAttributes.HttpFlavorValues.HTTP_2_0;
+    }
+    logger.log(Level.FINE, "unexpected http protocol version: {0}", protocolVersion);
+    return null;
   }
 
   @Override
@@ -67,6 +126,30 @@ final class ApacheHttpClientHttpAttributesGetter
 
   @Override
   public List<String> responseHeader(HttpRequest request, HttpResponse response, String name) {
-    return ApacheHttpClientUtils.getHeader(response, name);
+    return getHeader(response, name);
+  }
+
+  private static ProtocolVersion getVersion(HttpRequest request, @Nullable HttpResponse response) {
+    ProtocolVersion protocolVersion = request.getVersion();
+    if (protocolVersion == null && response != null) {
+      protocolVersion = response.getVersion();
+    }
+    return protocolVersion;
+  }
+
+  private static List<String> getHeader(MessageHeaders messageHeaders, String name) {
+    return headersToList(messageHeaders.getHeaders(name));
+  }
+
+  // minimize memory overhead by not using streams
+  private static List<String> headersToList(Header[] headers) {
+    if (headers.length == 0) {
+      return Collections.emptyList();
+    }
+    List<String> headersList = new ArrayList<>(headers.length);
+    for (Header header : headers) {
+      headersList.add(header.getValue());
+    }
+    return headersList;
   }
 }
