+++ instrumentation/apache-httpclient/apache-httpclient-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v5_0/ApacheHttpClientNetAttributesGetter.java
@@ -5,14 +5,19 @@
 
 package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v5_0;
 
+import static java.util.logging.Level.FINE;
+
 import io.opentelemetry.instrumentation.api.instrumenter.net.NetClientAttributesGetter;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+import java.util.logging.Logger;
 import javax.annotation.Nullable;
 import org.apache.hc.core5.http.HttpRequest;
 import org.apache.hc.core5.http.HttpResponse;
 
 final class ApacheHttpClientNetAttributesGetter
     implements NetClientAttributesGetter<HttpRequest, HttpResponse> {
+  private static final Logger logger =
+      Logger.getLogger(ApacheHttpClientNetAttributesGetter.class.getName());
 
   @Override
   public String transport(HttpRequest request, @Nullable HttpResponse response) {
@@ -22,12 +27,28 @@ final class ApacheHttpClientNetAttributesGetter
   @Override
   @Nullable
   public String peerName(HttpRequest request, @Nullable HttpResponse response) {
-    return ApacheHttpClientUtils.getPeerName(request);
+    return request.getAuthority().getHostName();
   }
 
   @Override
   public Integer peerPort(HttpRequest request, @Nullable HttpResponse response) {
-    return ApacheHttpClientUtils.getPeerPort(request);
+    int port = request.getAuthority().getPort();
+    if (port != -1) {
+      return port;
+    }
+    String scheme = request.getScheme();
+    if (scheme == null) {
+      return 80;
+    }
+    switch (scheme) {
+      case "http":
+        return 80;
+      case "https":
+        return 443;
+      default:
+        logger.log(FINE, "no default port mapping for scheme: {0}", scheme);
+        return null;
+    }
   }
 
   @Override
